refactor: combine QueryFormatter and QueryParser into QuerySerializer
- Replaced QueryFormatter and QueryParser with QuerySerializer with two
  nested public classes Formatter and Parser
- Updated query builders to use QuerySerializer
